backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic[email]==2.6.3
pydantic-settings==2.1.0
httpx==0.27.0
python-multipart==0.0.9